+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/planner/delegation/hive/DDLOperationConverter.java
@@ -394,7 +394,7 @@ public class DDLOperationConverter {
                             .collect(Collectors.toList());
             String constraintName = primaryKey.getConstraintName();
             if (constraintName == null) {
-                constraintName = "PK_" + pkCols.hashCode();
+                constraintName = pkCols.stream().collect(Collectors.joining("_", "PK_", ""));
             }
             uniqueConstraint = UniqueConstraint.primaryKey(constraintName, pkCols);
         }
