+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/planner/delegation/hive/HiveParserRexNodeConverter.java
@@ -602,7 +602,7 @@ public class HiveParserRexNodeConverter {
             }
             // type conversion for RHS
             RelNode rhsRel =
-                    HiveParser.addTypeConversions(
+                    HiveParserDMLHelper.addTypeConversions(
                             cluster.getRexBuilder(),
                             subQueryDesc.getRexSubQuery(),
                             Collections.singletonList(
