--- flink-connectors/flink-connector-hive/src/test/resources/query-test/queries/grouping_set.q
@@ -1,5 +1,0 @@
--- SORT_QUERY_RESULTS
-
-select x,y,grouping__id,sum(1) from foo group by x,y grouping sets ((x,y),(x));
-
-select x,y,grouping(x),sum(1) from foo group by x,y grouping sets ((x,y),(x));
