--- flink-connectors/flink-connector-hive/src/test/resources/query-test/results/order_by.q.out
@@ -1,11 +1,0 @@
-select x from foo order by x desc limit 1
-
-[+I[5]]
-
-select 'x' as key_new , split(value,',') as value_new from src ORDER BY key_new ASC, value_new[0] ASC limit 20
-
-[+I[x, [val1]], +I[x, [val2]], +I[x, [val3]]]
-
-select * from foo where cast(x as double)<=0 order by cast(x as double)
-
-[]
