--- flink-connectors/flink-connector-hive/src/test/resources/query-test/results/set_op.q.out
@@ -1,15 +1,0 @@
-select count(x) from foo union all select i from bar
-
-[+I[1], +I[1], +I[2], +I[5]]
-
-select x from foo union select i from bar
-
-[+I[1], +I[2], +I[3], +I[4], +I[5]]
-
-select i from bar except select x from foo
-
-[]
-
-select x from foo intersect select i from bar
-
-[+I[1], +I[2]]
