--- flink-connectors/flink-connector-hive/src/test/resources/query-test/results/temp_function.q.out
@@ -1,3 +1,0 @@
-select temp_abs(x) from foo
-
-[+I[1], +I[2], +I[3], +I[4], +I[5]]
