+++ flink-connectors/flink-connector-hive/src/test/resources/query-test/temp_function.q
@@ -1,3 +1,5 @@
 -- SORT_QUERY_RESULTS
 
 select temp_abs(x) from foo;
+
+[+I[1], +I[2], +I[3], +I[4], +I[5]]
